Update BlockDrillHead.java
i couldnt get intellij to work with susycore, so i couldnt test if it actually works.
this is the update, feel free to decline for any reason. Though if you you need the drill head to rotate, add a note into the book saying the drill head needs to face north.