Drone pad snow problem patch (#397)
* Update Drone behavior

It will only explode when hit SOLID blocks, blocks that have a collision shape, so 1 layer snow, string, grass won't cause the drone to explode.

Remove the age check so it will check for collision as soon as it flies, so full blocks that are put right above the pad will causse the drone to explode.

* Remove air check above pads

This check wont be needed because of the drone behaviors update.